modified in calculator_56.java for creating conflict

+++ Calculator_56.java
@@ -1,21 +1,21 @@
 public class Calculator_56 {
     // Add method
-    public double add(double a, double b) {
+    public float add(double a, double b) {
         return a + b;
     }
 
     // Subtract method
-    public double subtract(double a, double b) {
+    public float subtract(double a, double b) {
         return a - b;
     }
 
     // Multiply method
-    public double multiply(double a, double b) {
+    public float multiply(double a, double b) {
         return a * b;
     }
 
     // Divide method
-    public double divide(double a, double b) {
+    public float divide(double a, double b) {
         if (b == 0) {
             throw new IllegalArgumentException("Division by zero is not allowed.");
         }
